pretty much working and documented, but ugly
